Conflicts:
	EPMS.Web/EPMS.Web.csproj

+++ EPMS.Interfaces/Repository/IEmployeeRepository.cs
@@ -4,8 +4,9 @@ using EPMS.Models.ResponseModels;
 
 namespace EPMS.Interfaces.Repository
 {
-    public interface IEmployeeRepository : IBaseRepository<Employee, int>
+    public interface IEmployeeRepository : IBaseRepository<Employee, long>
     {
         EmployeeResponse GetAllEmployees(EmployeeSearchRequset employeeSearchRequset);
+
     }
 }
